Move StatusIcon to components/ui
