+++ airbyte-webapp/src/components/ui/StatusIcon/CircleLoader.module.scss
