+++ airbyte-webapp/src/components/ui/StatusIcon/StatusIcon.test.tsx
