+++ airbyte-webapp/src/components/ui/StatusIcon/StatusIcon.tsx
