+++ airbyte-webapp/src/views/Connection/ConnectionForm/components/CreateControls.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { FormattedMessage } from "react-intl";
 import styled from "styled-components";
 
-import { StatusIcon } from "components/common/StatusIcon";
 import { Button } from "components/ui/Button";
 import { Spinner } from "components/ui/Spinner";
+import { StatusIcon } from "components/ui/StatusIcon";
 
 interface CreateControlsProps {
   isSubmitting: boolean;
