+++ airbyte-webapp/src/views/Connector/ServiceForm/components/TestingConnectionSuccess.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { FormattedMessage } from "react-intl";
 import styled from "styled-components";
 
-import { StatusIcon } from "components/common/StatusIcon";
+import { StatusIcon } from "components/ui/StatusIcon";
 
 const LoadingContainer = styled.div`
   font-weight: 600;
